Move the `is_allowed_pg_catalog_rte` to allowed_functions.c
