+++ pg_diffix/query/allowed_functions.h
@@ -1,6 +1,8 @@
 #ifndef PG_DIFFIX_ALLOWED_FUNCTIONS_H
 #define PG_DIFFIX_ALLOWED_FUNCTIONS_H
 
+#include "nodes/bitmapset.h"
+
 /*
  * Returns whether the OID points to a function (or operator) allowed in defining buckets.
  */
@@ -27,5 +29,11 @@ extern bool is_substring_builtin(Oid funcoid);
  * allowed when called in `publish_untrusted` access level.
  */
 extern bool is_implicit_range_builtin_untrusted(Oid funcoid);
+
+/*
+ * Returns whether the OID points to a relation in the `pg_catalog`, for which accessing the
+ * `selected_cols` is allowed.
+ */
+extern bool is_allowed_pg_catalog_rte(Oid relation_oid, const Bitmapset *selected_cols);
 
 #endif /* PG_DIFFIX_ALLOWED_FUNCTIONS_H */
